refactor: rename the table file appropriately
